Refactor control builtins
